pyproject.toml
[project]
name = "CodeFitter"
version = "0.1.0"
description = "Code Fitter, 代码钳工，命令行编码智能体，半自动+精细编码"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "prompt-toolkit>=3.0.52",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "setuptools>=80.9.0",
]

[project.scripts]
fitter = "fitter.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
